Implement additional alternate file directory
This introduced a proof of concept implementation for an additional
directory which contains alternate files to be linked into ~. The
default path is ~/.yadm/alt. A file ~/.yadm/alt/foo## would be linked
to ~/foo.

The directory is optional and can be overriden using the configuration
option yadm.alt, which is a path relative to ~ (e.g. .yadm/my-alt).
The command-line option --yadm-alt allows overriding the path with an
absolute path (e.g. ~/.yadm/my-alt).

Caveats: This feature is hardly tested and does not come with any tests
or documentation. It will break if yadm.alt or --yadm-alt are not
normalized paths, but may break in many other cases, too (e.g. wildcards
in path names, esoteric symlink setups or just because of bugs).