Fix Makefile portability
OBS *among others* need to copy files from the build folder to the
package folder.
With the old version, that wasn't possible, as it would try to install the software in the worker folder,
of course something denied on public instances.
Adding $(DESTDIR) before all paths ensure that you can install to another folder